move project to separate file
